requirements.txt
Flask==3.0.*
SQLAlchemy==2.0.*
PyMySQL==1.1.*
python-dotenv==1.0.*
tabulate==0.9.*
tzdata